# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Danh sách các API để làm việc với Front End",
    "version": "1.0.0",
    "title": "API CỦA LPE"
  },
  "host": "localhost:8000",
  "basePath": "/api",
...
  "schemes": ["http"],
  "paths": {
...
    "/auth/change-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Đổi mật khẩu",
        "description": "",
        "operationId": "changePassword",
        "produces": ["application/json", "application/xml"],
        "parameters": [
          {
            "name": "Data",
            "in": "body",
            "description": "Gửi mật khẩu",
            "required": true,
            "type": "string",
            "schema": {
              "example": {
                "id": "60e6beddbeefa54ff879b36a",
                "newPassword": "123456@",
                "oldPassword": "manhdat10998@"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Đổi mật khẩu thành công",
            "schema": { "type": "string" }
          },
          "400": { "description": "Không tìm thấy người dùng" }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Quên mật khẩu (tạo Token)",
        "description": "",
        "operationId": "forgotPassword",
        "produces": ["application/json", "application/xml"],
        "parameters": [
          {
            "name": "Data",
            "in": "body",
            "description": "Gửi Email",
            "required": true,
            "type": "string",
            "schema": {
              "example": {
                "email": "vomanhdat10998@gmail.com"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công"
          },
          "400": { "description": "Email không tồn tại" }
        }
      }
    },
    "/auth/verify-forgot-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Quên mật khẩu (Xác thực Token)",
        "description": "",
        "operationId": "verifyForgotPassword",
        "produces": ["application/json", "application/xml"],
        "parameters": [
          {
            "name": "password",
            "in": "body",
            "description": "Gửi mật khẩu mới",
            "required": true,
            "type": "string",
            "schema": {
              "example": {
                "password": "123456789"
              }
            }
          },
          {
            "name": "token",
            "in": "headers",
            "description": "accessToken",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Đổi mật khẩu thành công!!!"
          },
          "400": { "description": "Error Occur" }
        }
      }
    },
    "/auth/verify-sign-up": {
      "post": {
        "tags": ["Auth"],
        "summary": "Đăng ký (Xác thực OTP)",
        "description": "",
        "operationId": "verifySignUp",
        "produces": ["application/json", "application/xml"],
        "parameters": [
          {
            "name": "otp",
            "in": "body",
            "description": "Gửi OTP",
            "required": true,
            "type": "string",
            "schema": {
              "example": {
                "otp": "141500"
              }
            }
          },
          {
            "name": "token",
            "in": "headers",
            "description": "accessToken",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Đăng ký thành công"
          },
          "403": { "description": "Nhập sai OTP" }
        }
      }
    }
  },
...
}